backend/requirements.txt
python-dotenv==1.0.1
langchain-community==0.3.2
langchain-openai==0.2.3
langchain-huggingface==0.1.0
langchain-groq==0.2.0
langchainhub==0.1.21
pypdf==4.2.0
faiss-cpu==1.9.0